+++ src/main/cpp/modules/Pcl.h
@@ -28,7 +28,7 @@ private:
     const Accessor* slnFlagsAccessor;
     const Accessor* sloFlagsAccessor;
 
-	size_t getIndex(size_t k, size_t l, size_t m) const;
+	size_t getIndex(long k, long l, long m) const;
 	uint16_t getValue(size_t index, int64_t olcFlags, int16_t slnFlags, int16_t sloFlags) const;
 	const Accessor& getSourceAccessor(Context& context, string variableName, string sourceSegmentId);
 	void setUpSourceAccessors(Context & context);
